style: 폴더 분리 & Test 분리
