+++ src/main/java/carracing/Main.java
@@ -1,5 +1,11 @@
 package carracing;
 
+import carracing.ui.InputView;
+import carracing.ui.OutputView;
+import carracing.controller.CarRacingGame;
+import carracing.domain.GameResult;
+import carracing.domain.RandomNumberGenerator;
+
 public class Main {
     public static void main(String[] args) {
         String carNumber = InputView.inputCarNumber();
