+++ src/main/java/carracing/domain/GameResult.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.domain;
 
 import java.util.List;
