+++ src/main/java/carracing/domain/NumberGenerator.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.domain;
 
 public interface NumberGenerator {
     int generateNumber();
