+++ src/main/java/carracing/ui/InputView.java
@@ -1,4 +1,4 @@
-package carracing;
+package carracing.ui;
 
 import java.util.Scanner;
